+++ src/Installers/InertiaInstaller.php
@@ -18,6 +18,11 @@ class InertiaInstaller implements InstallerContract
     {
         $this->output->write('Installing Inertia');
 
+        if (file_exists($welcome = resource_path('views/welcome.blade.php'))) {
+            unlink($welcome);
+            $this->output->write('.');
+        }
+
         copy(__DIR__.'/../../stubs/app.blade.php', resource_path('views').'/app.blade.php');
         $this->output->write('.');